KTO-1006: Code review fixes:
- EmptyResult.js replaced with already existing NotFound.js
- TulevaTarjonta card CSS fixed
